Adjust status helper for rare scenario
There's an edge case where a PT calls in and leaves a voicemail
indicating they have an appointment date. If we then set the appointment
date in the system, the status would have been 'Fundraising' but in
reality we would prefer it to say 'No Contact Made'.

This fix adjusts the priority of the statuses slightly that No Contact
Made will come up if we haven't talked to a user yet, regardless of
their appointment status.

Fixes issue #676.