Remove dropdown row UI
Clinics view used a dropdown row to show/hide additional info about a
given clinic. This UI was buggy. The dropdown row was implemented
through Bootstrap toggle and defined on a <tr> tag.

Bootstrap's toggle does not play well with the HTML <table> tag. A
workaround is possible but difficult. Instead, this commit takes the
three extra fields from the dropdown row and adds them as additional
columns within the table.

Before this commit, there were 5 columns of data and a chevron toggle
that would reveal 3 more columns of data in a dropdown row. Now
there is no chevron toggle, and the table displays 8 columns of data at
all times.