Improve validation messages for unique phone numbers
This change resolves a frustration point for CMs. Whenever they try to
register a patient whose phone number already exists in the system, they
will receive an error message stating the number is a duplicate. This
should not be a problem since CMs search for the number in the database
before creation. However, the search functionality is limited to the
CM's existing line, while the uniqueness check works across the entire
database.

For example, if the number 123-123-1231 exists on the DC line and the CM
is logged into the VA line, then they will recieve no results when
searching for 123-123-1231. This implies a new patient. Then when the CM
creates a new patient with 123-123-1231, they will get an error saying
the number already exists.

To address this issue, I added a custom validator that tells the CM when
the patient already exists on a different line.

Resolves: #825